+++ src/client.js
@@ -24,7 +24,8 @@ const defaultOpts = {
     accountUrl: null,
     backoffAttempts: 5,
     backoffMin: 5000,
-    backoffMax: 30000
+    backoffMax: 30000,
+    axiosOpts: {}
 };
 
 
@@ -39,6 +40,7 @@ const defaultOpts = {
  * @param {number} [opts.backoffAttempts] Maximum number of backoff attempts, default: `5`
  * @param {number} [opts.backoffMin] Minimum backoff attempt delay in milliseconds, default: `5000`
  * @param {number} [opts.backoffMax] Maximum backoff attempt delay in milliseconds, default: `30000`
+ * @param {object} [opts.axiosOpts] Default axios config (https://github.com/axios/axios#request-config)
  */
 
 class AcmeClient {
@@ -55,7 +57,7 @@ class AcmeClient {
             max: this.opts.backoffMax
         };
 
-        this.http = new HttpClient(this.opts.directoryUrl, this.opts.accountKey);
+        this.http = new HttpClient(this.opts.directoryUrl, this.opts.accountKey, this.opts.axiosOpts);
         this.api = new AcmeApi(this.http, this.opts.accountUrl);
     }